Share s3/SQS messager code
Instead of re-using the resolve hook to send maessages abstract out the common parts into a standalone messager component.
