+++ lib/hooks/create/project.js
@@ -1,11 +1,11 @@
 const { get } = require('lodash');
 const { autoResolved, withInspectorate } = require('../../flow/status');
 
-const Resolver = require('../resolve');
+const Messager = require('../../messager');
 
 module.exports = settings => {
   const { Project } = settings.models;
-  const resolver = Resolver(settings);
+  const messager = Messager(settings);
 
   return model => {
     const action = get(model, 'data.action');
@@ -26,7 +26,7 @@ module.exports = settings => {
             }
           });
       case 'grant':
-        return resolver({ data: { ...model.data, action: 'submit-draft' } })
+        return messager({ ...model.data, action: 'submit-draft' })
           .then(() => model.setStatus(withInspectorate.id));
     }
   };
