+++ lib/hooks/resolve/index.js
@@ -1,56 +1,10 @@
-const Queue = require('../../queue');
-const S3 = require('../../s3');
-const { TimeoutError } = require('../../errors');
-
-const { get } = require('lodash');
-
-// 30 second timeout
-const POLL_TIMEOUT = 30000;
-const POLL_INTERVAL = 300;
+const Messager = require('../../messager');
 
 module.exports = settings => {
-  const queue = Queue(settings.sqs);
-  const store = S3(settings.s3);
-
-  const { Changelog } = settings.models;
-
-  const pollChangelog = id => {
-    const endTime = Number(new Date()) + POLL_TIMEOUT;
-
-    const checkCondition = (resolve, reject) => {
-      console.log(`Polling changelog for: ${id}`);
-      Promise.resolve()
-        .then(() => Changelog.query().findById(id))
-        .then(model => {
-          if (model) {
-            return resolve(model);
-          } else if (Number(new Date()) < endTime) {
-            setTimeout(checkCondition, POLL_INTERVAL, resolve, reject);
-          } else {
-            return reject(new TimeoutError());
-          }
-        });
-    };
-    return new Promise(checkCondition);
-  };
-
-  if (settings.noDownstream) {
-    if (process.env.NODE_ENV === 'production') {
-      throw new Error('DO not set `noDownstream` configuration option in prod environments!');
-    }
-    return () => Promise.resolve();
-  }
+  const messager = Messager(settings);
 
   return model => {
-    return Promise.resolve()
-      .then(() => store(model.data))
-      .then(id => queue(id))
-      .then(response => pollChangelog(response.MessageId))
-      .then(changelogModel => {
-        const action = get(model, 'data.action');
-        if (action === 'create') {
-          return model.patch({ id: changelogModel.modelId });
-        }
-      });
+    return messager(model.data)
+      .then(changelogModel => model.patch({ id: changelogModel.modelId }));
   };
 };
